requirements.txt
alabaster==0.7.7
arrow==0.7.0
Babel==2.2.0
chai==1.1.1
docutils==0.12
gitdb==0.6.4
GitPython==1.0.2
Jinja2==2.8
MarkupSafe==0.23
nose==1.3.7
Pygments==2.1.3
python-dateutil==2.5.0
pytz==2015.7
six==1.10.0
smmap==0.9.0
snowballstemmer==1.2.1
Sphinx==1.3.6
sphinx-rtd-theme==0.1.9
